adopt for Python 3.6

+++ dereplicate.py
@@ -9,7 +9,7 @@ Technically, dereplication is OTU clustering with similarity = 1.0
 
 This script is slower than USEARCH 8.0, but does not have limitation on memory usage.
 
-Under windows, in order to use multiple thread, you need to run this program stand alone, 
+Under windows, in order to use multiple thread, you need to run this program stand alone,
 i.e. you can not summon this function using fast.py, but need to run dereplicate.py instead.
 
 Please feel free to contact me for any question.
@@ -59,7 +59,7 @@ def divide_seqs(total, thread_num):
     # Set break point for input sequences
     seqs_divide = []
     start = 0
-    size = total / thread_num
+    size = int(total / thread_num)
     for i in range(thread_num):
         seqs_divide.append([start, start + size])
         start += size
@@ -83,16 +83,16 @@ def main(Namespace):
     parser.add_argument('-t', '--thread', default = 1, help='Number of threads to be used.')
     parser.add_argument('-fast', default = "", help="Name of FAST style output file.")
     parser.add_argument('-sizeout', action = 'store_true', help='Specify to add a USEARCH style size label: ";szie=XXX"')
-    
+
     args = parser.parse_args(Namespace)
-    
+
     input_file = args.input
     output_name = args.output
     output_map = output_name + '.txt'
     output_fasta = output_name + '.fasta'
 
     thread = int(args.thread)
-    
+
     import time
     from lib import File_IO
     from multiprocessing import Process, Manager
@@ -100,7 +100,7 @@ def main(Namespace):
 
     print('Using %i threads ...' % thread)
     start = time.time()
-    
+
     input_file = input_file
     print('Loading %s ...' % input_file)
     seqs = File_IO.read_seqs(input_file)
@@ -114,13 +114,13 @@ def main(Namespace):
         # Separated seqs into pools
         print('Separating raw sequences into %d jobs ...' % thread)
         d = divide_seqs(seqs_num, thread)
-    
-        
+
+
         # Create shared list for store dereplicated dict and progress counter
         manager = Manager()
         derep_dict = manager.list([{}] * thread)
         count = manager.list([0] * thread)
-    
+
         print('Starting dereplicating ...')
         workers = []
         for i in range(thread):
@@ -128,24 +128,24 @@ def main(Namespace):
             workers.append(Process(target=dereplicate_worker,
                                    args=(seqs[current_range[0]:current_range[1]], derep_dict, i, count)))
         del seqs
-        
+
         print('Starting %i jobs ...' % thread)
         count_worker = 1
         for job in workers:
             job.start()
             print('Starting thread No. %i ...' % count_worker)
             count_worker += 1
-            
+
         job_alive = True
         while job_alive:
             time.sleep(0.01)
             job_alive = False
-            for job in workers:            
+            for job in workers:
                 if job.is_alive():
                     job_alive = True
             #progress = "Dereplicating: " + str(round(sum(count)/float(seqs_num)*100,2)) + "%" + "\r"
             #sys.stderr.write(progress)
-    
+
         for derep_worker in workers:
             derep_worker.join()
         print('Finished dereplicating.')
@@ -153,7 +153,7 @@ def main(Namespace):
 
     # Merged dereplicated dictionaries into a single dict
     sys.stderr.write('\n')
-    
+
     if thread > 1:
         sys.stderr.write('Merging %i dictionaries into one ...' % len(derep_dict))
         merged_dict = {}
@@ -166,7 +166,7 @@ def main(Namespace):
                 except KeyError:
                     merged_dict[key] = value
                 #sys.stderr.write('Merging %i sequence ...' % count + '\b' * 50,)
-            derep_dict[0] = ''  # Empty finished dictionary to free memory.  
+            derep_dict[0] = ''  # Empty finished dictionary to free memory.
     else:
         merged_dict = derep_dict
     print
@@ -176,8 +176,8 @@ def main(Namespace):
     end = time.time()
     print("Used time: " + str(end - start) + ' seconds.')
     print
-    
-  
+
+
     # Name the dereplicated group
     size_list = sorted([[len(merged_dict[i]), i] for i in merged_dict], reverse=True)
     count = 0
@@ -191,7 +191,7 @@ def main(Namespace):
     print('Writing dereplicated sequence and OTU map ...')
     output_seq_file = output_fasta
     with open(output_seq_file, 'w') as f:
-        if args.sizeout:        
+        if args.sizeout:
             for element in size_list:
                 output_label = element[2] + ";size=" + str(element[0])
                 f.write('>%s\n' % output_label)
@@ -202,7 +202,7 @@ def main(Namespace):
                 output_label = element[2]
                 f.write('>%s\n' % output_label)
                 f.write('%s\n' % element[1])
-                
+
     print('%s contains dereplicated sequences.' % output_fasta)
 
     # Output Qiime style map
@@ -211,18 +211,18 @@ def main(Namespace):
             name_list = merged_dict[element[1]]
             f.write('%s\t%s\n' % (element[2], '\t'.join(name_list)))  # Use the last element as group name
     print('%s contains an OTU map for dereplicated sequences.' % output_map)
-    
+
     # Generate FAST style derep output file (a single file with sample names, counts, and dereplicated sequences)
     if args.fast != "":
         fast_file = args.fast
-        
+
         fast_dict = {}
-        
+
         for element in size_list:
-            
+
             fast_dict[element[2]] = {} # Crearte a new dict for current derep unit
             fast_dict[element[2]]['seq'] = element[1] # Save dereplicated sequence
-            
+
             sample_dict = {} # Create a dict for sample sequence count
             name_list = merged_dict[element[1]]
             for sample in name_list:
@@ -231,9 +231,9 @@ def main(Namespace):
                     sample_dict[current_sample] += 1
                 except KeyError:
                     sample_dict[current_sample] = 1
-            
+
             fast_dict[element[2]]['sample'] = sample_dict
-        
+
         import json
         json.dump(fast_dict, open(fast_file, "wb"))
